+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Allow user to define their own escape function, if you want to generate JSON for example
 - Add `end` argument to the `truncate` filter to override the default ellipsis
 - Add a `group_by` filter
+- Add the `~` operator to concat strings
 
 
 ## 0.11.7 (2018-04-24)